Cleaned up GUI code a bit, improved MVC pattern
I moved the MenuBar content to its own class, separated various parts of the GUI setup that were similar into functions, and improved the Model-View-Controller pattern by a bit for cleanliness and understandability. The startup list is now in the controller so that it can make calls to the API if desired.